글작성, 글수정권한체크, 글삭제권한체크 시 세션의 정보 사용

+++ src/com/sbs/example/jspCommunity/controller/UsrArticleController.java
@@ -76,7 +76,7 @@ public class UsrArticleController {
 			return "common/redirect";
 		}
 
-		int memberId = Integer.parseInt(req.getParameter("memberId"));
+		int memberId = (int)session.getAttribute("loginedMemberId");
 		int boardId = Integer.parseInt(req.getParameter("boardId"));
 		String title = req.getParameter("title");
 		String body = req.getParameter("body");
@@ -113,6 +113,14 @@ public class UsrArticleController {
 			req.setAttribute("historyBack", true);
 			return "common/redirect";
 		}
+		
+		int loginedMemberId = (int)session.getAttribute("loginedMemberId");
+
+		if (article.getMemberId() != loginedMemberId) {
+			req.setAttribute("alertMsg", id + "번 게시물에 대한 권한이 없습니다.");
+			req.setAttribute("historyBack", true);
+			return "common/redirect";
+		}
 
 		articleService.delete(id);
 
@@ -139,6 +147,14 @@ public class UsrArticleController {
 
 		if (article == null) {
 			req.setAttribute("alertMsg", id + "번 게시물은 존재하지 않습니다.");
+			req.setAttribute("historyBack", true);
+			return "common/redirect";
+		}
+		
+		int loginedMemberId = (int)session.getAttribute("loginedMemberId");
+
+		if (article.getMemberId() != loginedMemberId) {
+			req.setAttribute("alertMsg", id + "번 게시물에 대한 권한이 없습니다.");
 			req.setAttribute("historyBack", true);
 			return "common/redirect";
 		}
@@ -171,9 +187,9 @@ public class UsrArticleController {
 			return "common/redirect";
 		}
 
-		int memberId = Integer.parseInt(req.getParameter("memberId"));
+		int loginedMemberId = (int)session.getAttribute("loginedMemberId");
 
-		if (article.getMemberId() != memberId) {
+		if (article.getMemberId() != loginedMemberId) {
 			req.setAttribute("alertMsg", id + "번 게시물에 대한 권한이 없습니다.");
 			req.setAttribute("historyBack", true);
 			return "common/redirect";
